requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.0
python-dotenv==1.0.0
openai==1.10.0